+++ back/api/param.go
@@ -24,7 +24,7 @@ func GetParam(w http.ResponseWriter, r *http.Request, isFrontRequest bool,
 	return appdata.RenderAPIorUI(w, r, isFrontRequest, true, true, http.StatusOK, "user params retrieved", userParams)
 }
 
-func cleanStringList(stringList string) string {
+func cleanStringList(stringList string, minStringLength int) string {
 	var list []string
 	var cleanedString string
 	var cleanedStringResult string = ""
@@ -32,7 +32,7 @@ func cleanStringList(stringList string) string {
 	for _, element := range list {
 		// remove all " " and "," from beginning and end
 		cleanedString = strings.Trim(element, " ,")
-		if len(cleanedString) > 1 {
+		if len(cleanedString) >= minStringLength {
 			if cleanedStringResult != "" {
 				cleanedStringResult += ","
 			}
@@ -41,7 +41,7 @@ func cleanStringList(stringList string) string {
 	}
 	return cleanedStringResult
 }
-func postParam(w http.ResponseWriter, r *http.Request, isFrontRequest bool, paramName string, paramInfo string) *appdata.HttpStruct {
+func postParam(w http.ResponseWriter, r *http.Request, isFrontRequest bool, paramName string, paramInfo string, minStringLength int) *appdata.HttpStruct {
 	param := &appdata.Param{}
 	if err := render.Bind(r, param); err != nil {
 		fmt.Printf("error: %v\n", err.Error())
@@ -50,8 +50,9 @@ func postParam(w http.ResponseWriter, r *http.Request, isFrontRequest bool, para
 	userContext := r.Context().Value(appdata.ContextUserKey).(*appdata.UserRequest)
 	param.GofiID = userContext.GofiID
 	param.ParamName = paramName
-	param.ParamJSONstringData = cleanStringList(param.ParamJSONstringData)
+	param.ParamJSONstringData = cleanStringList(param.ParamJSONstringData, minStringLength)
 	param.ParamInfo = paramInfo
+	// fmt.Printf("param: %#v\n", param)
 	_, err := sqlite.InsertRowInParam(r.Context(), appdata.DB, param)
 	if err != nil { // Always check errors even if they should not happen.
 		fmt.Printf("error: %v\n", err.Error())
@@ -62,12 +63,34 @@ func postParam(w http.ResponseWriter, r *http.Request, isFrontRequest bool, para
 func PostParamAccount(w http.ResponseWriter, r *http.Request, isFrontRequest bool) *appdata.HttpStruct {
 	return postParam(w, r, isFrontRequest,
 		"accountList",
-		"Liste des comptes (séparer par des , sans espaces)")
+		"Liste des comptes (séparer par des , sans espaces)",
+		2) // minStringLength to handle the value
 }
 func PostParamCategoryRendering(w http.ResponseWriter, r *http.Request, isFrontRequest bool) *appdata.HttpStruct {
 	return postParam(w, r, isFrontRequest,
 		"categoryRendering",
-		"Affichage des catégories: icons | names")
+		"Affichage des catégories: icons | names",
+		1) // minStringLength to handle the value
+}
+func PostParamOnboardingCheckList(w http.ResponseWriter, r *http.Request, isFrontRequest bool) *appdata.HttpStruct {
+	return postParam(w, r, isFrontRequest,
+		"onboardingCheckList",
+		"Liste des étapes (séparer par des , sans espaces)",
+		1) // minStringLength to handle the value
+}
+func UpdateParamCheckList(w http.ResponseWriter, r *http.Request, isFrontRequest bool, 
+	gofiID int, onboardingCheckListSingleString string) *appdata.HttpStruct {
+	param := &appdata.Param{}
+	param.GofiID = gofiID
+	param.ParamName = "onboardingCheckList"
+	param.ParamJSONstringData = onboardingCheckListSingleString
+	param.ParamInfo = "Liste des étapes (séparer par des , sans espaces)"
+	_, err := sqlite.InsertRowInParam(r.Context(), appdata.DB, param)
+	if err != nil { // Always check errors even if they should not happen.
+		fmt.Printf("error: %v\n", err.Error())
+		return appdata.RenderAPIorUI(w, r, isFrontRequest, true, false, http.StatusInternalServerError, "server error", "")
+	}
+	return appdata.RenderAPIorUI(w, r, isFrontRequest, false, false, http.StatusOK, "onboardingCheckList updated", "")
 }
 
 func GetCategoryIcon(w http.ResponseWriter, r *http.Request, isFrontRequest bool, categoryNameFuncParam string, cd *appdata.CategoryDetails) *appdata.HttpStruct {