+++ back/routes/routes.go
@@ -116,6 +116,7 @@ func (s *Server) MountBackHandlers() {
 				r.Patch("/category/in-use", func(w http.ResponseWriter, r *http.Request) { api.PatchParamCategoryInUse(w, r, false) })
 				r.Patch("/category/order", func(w http.ResponseWriter, r *http.Request) { api.PatchParamCategoryOrder(w, r, false) })
 				r.Post("/category-rendering", func(w http.ResponseWriter, r *http.Request) { api.PostParamCategoryRendering(w, r, false) })
+				r.Post("/onboardingchecklist", func(w http.ResponseWriter, r *http.Request) { api.PostParamOnboardingCheckList(w, r, false) })
 			})
 			// RECORDS
 			r.Route("/record", func(r chi.Router) {
@@ -168,12 +169,25 @@ func (s *Server) MountFrontHandlers() {
 			// PRIVATE
 			r.Use(appmiddleware.AuthenticatedUserOnly)
 
+			// CHECKLIST
+			r.Route("/checklist", func(r chi.Router) {
+				r.Get("/", front.Checklist)
+				r.Get("/1", front.Checklist1)
+				r.Get("/2", front.Checklist2)
+				r.Get("/3", front.Checklist3)
+				r.Get("/4", front.Checklist4)
+				r.Get("/5", front.Checklist5)
+				r.Get("/6", front.Checklist6)
+				r.Get("/7", front.Checklist7)
+				r.Get("/8", front.Checklist8)
+			})
 			// PARAMS
 			r.Route("/param", func(r chi.Router) {
 				r.Get("/", front.GetParam)
 				r.Get("/account", front.GetParamAccount)
 				r.Post("/account", front.PostParamAccount)
 				r.Post("/category-rendering", front.PostParamCategoryRendering)
+				r.Post("/onboardingchecklist", front.PostParamOnboardingCheckList)
 				r.Get("/category", front.GetParamCategory)
 				r.Put("/category", front.PutParamCategory)
 				r.Patch("/category/in-use", front.PatchParamCategoryInUse)