+++ CHANGELOG.md
@@ -3,9 +3,13 @@
 ## v1 rework with go-chi
 ### Doing
 
-
 ### Done
-- 2025-01-18 : chore: update go modules and go main version
+- 2025-01-19 : Branch onboarding checklist merged
+    - feat: add a 8 steps checklist to discover GOFI basics
+- 2025-01-18 : 
+    - chore: update go modules and go main version
+    - NewBranch onboarding checklist
+        - feat: add onboardingCheckList param
 - 2025-01-12 : feat: force new login on client change 
 - 2024-12-23 : 
     - fix: test param and first accounts on user creation